[fix] UserRepository의 UserId를 Long -> Integer로 변경

+++ BE/src/main/java/com/FTIsland/BE/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     //Optional<User> findByEmail(String email);
     List<Optional<User>> findByInputId(String inputId);
@@ -18,7 +18,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     //Optional<User> findByRefreshToken(String refreshToken);
 
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 
     /**
      * 소셜 타입과 소셜의 식별값으로 회원 찾는 메소드
